Use ints instead of uint32/int64 to cleanup the code...
While the uints are more precise, the extra work required to manipulate
the unsigned ints just obscures the code without adding any benefit as
far as protection from negative values.

+++ dierollerpkg/dieroll_test.go
@@ -10,19 +10,19 @@ func TestConstructorAndGetters(t *testing.T) {
 	}
 	var dr DieRoll
 	dr = *NewDieRoll(10, 20, "", 5)
-	if dr.Dice() != uint32(10) {
+	if dr.Dice() != 10 {
 		t.Error("Expected dr.Dice() to be 10, got ", dr.Dice())
 	}
-	if dr.Sides() != uint32(20) {
+	if dr.Sides() != 20 {
 		t.Error("Expected dr.Sides() to be 20, got ", dr.Sides())
 	}
 	if dr.Modifier().ModType != DieModifierTypeNull {
 		t.Error("Expected dr.Modifier().ModType to be DieModifierTypeNull, got ", dr.Modifier().ModType)
 	}
-	if dr.Modifier().Amount != uint32(0) {
+	if dr.Modifier().Amount != 0 {
 		t.Error("Expected dr.Modifier().Amount to be 0, got ", dr.Modifier().Amount)
 	}
-	if dr.Keep() != uint32(5) {
+	if dr.Keep() != 5 {
 		t.Error("Expected dr.Keep() to be 5, got ", dr.Keep())
 	}
 	if len(dr.History()) != 0 {
@@ -33,13 +33,13 @@ func TestConstructorAndGetters(t *testing.T) {
 func TestRoll(t *testing.T) {
 	dr := *NewDieRoll(10, 20, "", 5)
 	r := dr.Roll()
-	if r.Result < int64(5) || r.Result > int64(100) {
+	if r.Result < 5 || r.Result > 100 {
 		t.Error("Expected r.Result to be between 5 and 100 inclusive, got ", r.Result)
 	}
-	if uint32(len(r.Rolls)) != dr.Keep() {
+	if len(r.Rolls) != dr.Keep() {
 		t.Errorf("Expected r.Rolls to have %d entries, got ", dr.Keep(), len(r.Rolls))
 	}
-	if int64(r.Rolls[0]+r.Rolls[1]+r.Rolls[2]+r.Rolls[3]+r.Rolls[4]) != r.Result {
+	if r.Rolls[0]+r.Rolls[1]+r.Rolls[2]+r.Rolls[3]+r.Rolls[4] != r.Result {
 		t.Errorf("Expected r.Result to be the sum of %d,%d,%d,%d, & %d, got ", r.Rolls[0], r.Rolls[1], r.Rolls[2], r.Rolls[3], r.Rolls[4], r.Result)
 	}
 }
